requirement.txt
absl-py==1.4.0
addict==2.4.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
audioread==3.0.0
av==10.0.0
backcall==0.2.0
basicsr==1.3.5
bitarray==2.7.3
cachetools==5.3.0
certifi==2022.12.7
click==8.1.3
clip==1.0
configargparse==1.5.3
cpbd==1.0.7
cycler==0.11.0
cython==0.29.33
decorator==5.1.1
dlib==19.24.1
docker-pycreds==0.4.0
dominate==2.7.0
einops==0.6.0
exceptiongroup==1.1.1
executing==1.2.0
face-alignment==1.3.5
facexlib==0.3.0
fairseq==0.12.2
fastdtw==0.3.4
filterpy==1.4.5
flatbuffers==23.3.3
ftfy==6.1.1
gast==0.4.0
gitdb==4.0.10
gitpython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
hydra-core==1.0.7
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
install==1.3.5
interval==1.0.0
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
kornia==0.6.10
lazy-loader==0.1
libclang==15.0.6.1
librosa==0.7.2
llvmlite==0.31.0
lmdb==1.4.0
lws==1.2.7
lxml==4.9.2
markdown==3.4.1
markupsafe==2.1.2
matplotlib==3.4.3
matplotlib-inline==0.1.6
msgpack==1.0.5
networkx==3.0
ninja==1.11.1
numba==0.48.0
numpy==1.23.4
nvdiffrast==0.3.0
oauthlib==3.2.2
omegaconf==2.0.6
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opentsne==0.7.1
opt-einsum==3.3.0
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
piq==0.7.1
pluggy==1.0.0
pooch==1.6.0
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
pygments==2.14.0
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2022.7.1
pywavelets==1.4.1
realesrgan==0.3.0
regex==2023.3.23
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.9
sacrebleu==2.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
sentry-sdk==1.17.0
setproctitle==1.3.2
setuptools==59.5.0
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.5
stack-data==0.6.2
tabulate==0.9.0
tb-nightly==2.13.0a20230423
tensorboard==2.11.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardx==2.6
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.3.15
tomli==2.0.1
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchtext==0.11.0
torchvision==0.11.0+cu111
traitlets==5.9.0
trimesh==3.20.2
wandb==0.14.0
wcwidth==0.2.6
werkzeug==2.2.3
wheel==0.40.0
wrapt==1.15.0
yacs==0.1.8
yapf==0.33.0
zipp==3.15.0